interpolação

+++ Ray.c
@@ -29,8 +29,7 @@ void Ray_setDirection(int dx, int dy, int dz) {
 // raio é uma reta parametrica definida por um vetor de origem e um vetor direção
 // (ox, oy, oz) + t*(dx, dy, dz)
 // retorna o indice do volume dado um t
-// @TODO: interpolar valor (para o caso de t nao ser inteiro)
-int Ray_trace(double t) {
+int Ray_trace(int t) {
 	int i, j, k;
 	i = _origin[0] + t*_direction[0];
 	j = _origin[1] + t*_direction[1];
